ICacheKeyGenerator interface
