+++ Nancy.LightningCache/CacheKey/ICacheKeyGenerator.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Nancy.LightningCache.CacheKey
+{
+    /// <summary>
+    /// CacheKeyGenerator meant to be consumed by Nancy.LightningCache
+    /// </summary>
+    public interface ICacheKeyGenerator
+    {
+        string Get(Request request);
+    }
+}
