+++ Nancy.LightningCache/Nancy.LightningCache.csproj
@@ -49,6 +49,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CacheKey\ICacheKeyGenerator.cs" />
+    <Compile Include="CacheKey\DefaultCacheKeyGenerator.cs" />
     <Compile Include="Extensions\NancyBootstrapperExtensions.cs" />
     <Compile Include="Extensions\NegotiatorExtensions.cs" />
     <Compile Include="Projection\CachedResponse.cs" />
